refactor: 사다리를 출력하기 위한 구성 요소를 view 패키지로 이동
